Appraisal Module done

+++ backend/system/models.py
@@ -185,11 +185,9 @@ class Employee(AbstractUser):
         related_name="employees"
     )
 
-    basic_salary = models.DecimalField(
-        max_digits=10,
-        decimal_places=2,
+    basic_salary = models.IntegerField(
         null=True, blank=True,
-        default=0.00
+        default=0
     )
 
     reporting_manager = models.ForeignKey(
